Support for programatically naming exports
